src/models/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.1.0",
    "google-genai>=1.39.1",
    "google-cloud-storage>=2.18.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "openai>=1.95.1",
    "openai-agents>=0.3.2",
    "pandas>=2.3.3",
    "pyarrow>=17.0.0",
    "tqdm>=4.67.1",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "black>=24.1.1",
    "flake8>=7.0.0",

]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

# Make this pyproject.toml NON‑INSTALLABLE so it does not interfere
[tool.setuptools]
packages = []
package-dir = {"" = "."}
